Move assertions into shared utility module.
